+++ tests/fixtures/event.fixture.js
@@ -1,7 +1,0 @@
-module.exports = {
-  'startDate'  : '20150704T190000',
-  'endDate'    : '20150704T210000',
-  'title'       : 'Fourth of July Fireworks',
-  'location'    : 'Battery Park City, New York, NY',
-  'description' : 'Hello, world!'
-};