+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## v1.3.5
 
-- removes bower support
+- adds full recurrence support
+- adds vanilla js/core calendar library
+- upgrades FileSaver to address Safari support issues
+- fixes Yahoo end time issues and addresses Outlook calendar upgrade
+- corrects peer dependency issues
 
 ## v1.3.4